+++ test/adapters/mysql.rb
@@ -1,1 +1,1 @@
-ActiveRecord::Base.establish_connection adapter: "mysql2", database: "groupdate_test", username: "root"
+ActiveRecord::Base.establish_connection adapter: "mysql2", database: "groupdate_test"